Set the consistency threshold for Cartpole to be 0.01 instead of 20.
Added a meaningful error message when the plan to simulate is empry.

Made sure the Cartpole meta model creates valid PDDL+ states, where the fluent names are always tuples, even if it is a tuple with a single element.
There's a python quirk here, to initialize a tuple with a single element that is a string, it needs to look like this ("mystring",) and not ("mystring").